Increase Precision
Swap float32 to float64 because I occasionally get errors based on this.

+++ jenks.pyx
@@ -5,8 +5,8 @@ cimport cython
 # "ctypedef" assigns a corresponding compile-time type to DTYPE_t. For
 # every type in the numpy module there's a corresponding compile-time
 # type with a _t-suffix.
-DTYPE = np.float32
-ctypedef np.float32_t DTYPE_t
+DTYPE = np.float64
+ctypedef np.float64_t DTYPE_t
 
 
 @cython.boundscheck(False)
